// portable-deployment/source/src/types.ts
export type DownloadItem = { name: string; version: string; description: string; updated: string; url: string; size?: string; checksum?: { md5?: string; sha256?: string }; diff?: { url: string; size: string } }
export type DownloadData = Record<'android' | 'windows' | 'ios' | 'macos', DownloadItem[]>